Add some comments

+++ NOpenType/RendererOptions.cs
@@ -3,8 +3,17 @@ using System;
 
 namespace NRasterizer
 {
+    /// <summary>
+    /// Configuration options for controlling how text is rendered.
+    /// </summary>
     public class RendererOptions
     {
-        public int FontSize { get; set; } = 10;
+        /// <summary>
+        /// Gets or sets the size of the font in point
+        /// </summary>
+        /// <value>
+        /// The size of the font.
+        /// </value>
+        public float FontSize { get; set; } = 10;
    }
 }
